+++ src/main/java/in/koreatech/koin/domain/kakaobot/dto/KakaoDiningRequest.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.integration.kakao.dto;
+package in.koreatech.koin.domain.kakaobot.dto;
 
 import in.koreatech.koin.domain.dining.model.DiningType;
 import io.swagger.v3.oas.annotations.Hidden;
