--- src/main/java/in/koreatech/koin/integration/kakao/model/KakaoSkillRequestParser.java
@@ -1,8 +1,0 @@
-package in.koreatech.koin.integration.kakao.model;
-
-import in.koreatech.koin.integration.kakao.exception.KakaoApiException;
-
-public interface KakaoSkillRequestParser<T, R> {
-
-    R parse(T request) throws KakaoApiException;
-}
